requirements.txt
asgiref==3.6.0
attrs==23.1.0
CacheControl==0.13.0
cachetools==5.3.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
Django==4.0.4
django-cors-headers==3.14.0
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docopt==0.6.2
drf-spectacular==0.26.2
drf-spectacular-sidecar==2023.6.1
drf-yasg==1.21.6
firebase-admin==6.1.0
google-api-core==2.11.0
google-api-python-client==2.88.0
google-auth==2.19.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
grpcio==1.54.2
grpcio-status==1.54.2
gunicorn==20.1.0
httplib2==0.22.0
idna==3.4
inflection==0.5.1
jsonschema==4.17.3
msgpack==1.0.5
packaging==23.1
Pillow==9.5.0
pipreqs==0.4.11
proto-plus==1.22.2
protobuf==4.23.2
psycopg2-binary==2.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-decouple==3.8
pytz==2023.3
PyYAML==6.0
requests==2.28.2
rsa==4.9
six==1.16.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.15
whitenoise==6.2.0
yarg==0.1.9
